+++ compiler/rustc_error_messages/src/lib.rs
@@ -20,7 +20,7 @@ use tracing::{instrument, trace};
 #[cfg(not(parallel_compiler))]
 use std::cell::LazyCell as Lazy;
 #[cfg(parallel_compiler)]
-use std::lazy::SyncLazy as Lazy;
+use std::sync::LazyLock as Lazy;
 
 #[cfg(parallel_compiler)]
 use intl_memoizer::concurrent::IntlLangMemoizer;
