+++ src/tools/clippy/tests/test_utils/mod.rs
@@ -1,9 +1,9 @@
 #![allow(dead_code)] // see https://github.com/rust-lang/rust/issues/46379
 
-use std::lazy::SyncLazy;
 use std::path::PathBuf;
+use std::sync::LazyLock;
 
-pub static CARGO_CLIPPY_PATH: SyncLazy<PathBuf> = SyncLazy::new(|| {
+pub static CARGO_CLIPPY_PATH: LazyLock<PathBuf> = LazyLock::new(|| {
     let mut path = std::env::current_exe().unwrap();
     assert!(path.pop()); // deps
     path.set_file_name("cargo-clippy");
